create a custom hook for data fetch and update in the store
