+++ src/hooks/useNowPlayingMovies.js
@@ -1,0 +1,21 @@
+import { useDispatch } from "react-redux";
+import { API_OPTIONS, NOW_PLAYING_MOVIES_URL } from "../utils/constants";
+import { addNowPlayingMovies } from "../utils/movieSlice";
+import { useEffect } from "react";
+
+const useNowPlayingMovies = () => {
+  const dispatch = useDispatch();
+
+  const getNowPlayingMovies = async () => {
+    const data = await fetch(NOW_PLAYING_MOVIES_URL, API_OPTIONS);
+    const json = await data.json();
+    // console.log(json.results);
+    dispatch(addNowPlayingMovies(json.results));
+  };
+
+  useEffect(() => {
+    getNowPlayingMovies();
+  }, []);
+};
+
+export default useNowPlayingMovies;
